Adding role for installing CLI tools
This is the first draft of the role to install CLI tools. It pulls configs from
drduh/config and adds them to their relevant locations in the file system. It
also starts the DNS/Proxy services, adds them as login items.

todo: Add system commands to change DNS settings to point to these services