Fix: ignore errors in homebrew role so we don't get derailed when cask misbehaves
Because: if you're establishing a baseline on an already active system, some of the apps may be installed. Cask is not totally idempotent, so sometimes this is fine, but other times it results in tasks failing. There's no sense having the whole playbook stop because cask tried to install an app that was already there and got pissed.