Make the confdir the homebrew etc directory, not the system one.
Because: the default configuration here results in privoxy shutting off the internet entirely for our web browsers.

Todo: It does appear that something is still awry in our privoxy configs. While the internet is working again, navigating to config.privoxy.org leads to a page that claims privoxy is not being used. So further debugging is needed. This isn't mission-critical: our hosts file is going to shut down most ads, so the proxy is kinda just icing on top.