Disabling Privoxy  (for the time being)
* Fix: ignore errors in homebrew role so we don't get derailed when cask misbehaves

Because: if you're establishing a baseline on an already active system, some of the apps may be installed. Cask is not totally idempotent, so sometimes this is fine, but other times it results in tasks failing. There's no sense having the whole playbook stop because cask tried to install an app that was already there and got pissed.

* Fix: use flat so that ansible doesn't create extraneous directory layers

Because: All we want here is to have .curlrc in our home directory.

* Make the confdir the homebrew etc directory, not the system one.

Because: the default configuration here results in privoxy shutting off the internet entirely for our web browsers.

Todo: It does appear that something is still awry in our privoxy configs. While the internet is working again, navigating to config.privoxy.org leads to a page that claims privoxy is not being used. So further debugging is needed. This isn't mission-critical: our hosts file is going to shut down most ads, so the proxy is kinda just icing on top.

* Ignore files corresponding to installed roles.

* Update: include bitwarden-cli in our installed packages

* Temp: Don't configure Privoxy

Because: the fix I thought I'd made doesn't seem to be working. So we
need to do some experimenting with Privoxy to get it to work. In the
meantime, it's not a particularly important part of our setup, so
we're just disabling it.

* Fix: Install script syntax

Because: Removing some syntax from the install script that I think was
copied over from Travis but isn't actual shell syntax. Also, I don't
think we need the ansible config line, but we are going to test that.

* Test: comment out a line I don't think we need.